# Reduce allocation in Index/Range.ToString