@stephentoub 2019-01-02
> Nit: should we have a debug assert after the calls to TryFormat? I could go either way on this.

Thanks, Levi.  I considered that, too, but was on the fence and ended up removing them to avoid the clutter.  Since you also suggested it, I'll add them back.